feat: migrate election_with_member_votes_table.spec
